bugfix: 拆分 system_mgmt NATS handler
关联 Issue #3461：system_mgmt/nats_api.py 集中承载全部 NATS handler，职责混杂导致维护风险过高。

变更内容：

- 将 handler 按认证、用户、通道、权限、登录、OTP、微信、设置、审计等职责拆分到 apps/system_mgmt/nats/

- 保留 apps.system_mgmt.nats_api 兼容导出层，避免旧调用方导入路径失效

- apps.py 改为导入新 nats package 统一触发 handler 注册
