+++ server/apps/system_mgmt/apps.py
@@ -4,6 +4,5 @@ from django.apps import AppConfig
 class HandleConfig(AppConfig):
     name = "apps.system_mgmt"
 
-    #
     def ready(self):
-        import apps.system_mgmt.nats_api  # noqa
+        import apps.system_mgmt.nats  # noqa
