+++ server/apps/system_mgmt/nats_api.py
@@ -1,1979 +1,26 @@
-import base64
-import io
-import json
-import os
-import time
-from datetime import timedelta
-from uuid import uuid4
+# flake8: noqa
+"""Compatibility exports for system_mgmt NATS handlers.
 
-import jwt
-import pyotp
-import qrcode
-from django.contrib.auth.hashers import check_password, make_password
-from django.core.cache import cache
-from django.db import transaction
-from django.db.models import Q
-from django.utils import timezone
+Handlers live in apps.system_mgmt.nats and are imported here so older tests and
+callers using apps.system_mgmt.nats_api keep working.
+"""
 
-import nats_client
-from apps.core.constants import VERIFY_TOKEN_USER_NOT_FOUND_CODE, VERIFY_TOKEN_USER_NOT_FOUND_MESSAGE
-from apps.core.logger import system_mgmt_logger as logger
-from apps.core.utils.loader import LanguageLoader
-from apps.core.utils.permission_cache import clear_token_info_cache, clear_users_permission_cache, get_cached_token_info, set_cached_token_info
-from apps.system_mgmt.guest_menus import CMDB_MENUS, MONITOR_MENUS, OPSPILOT_GUEST_MENUS
-from apps.system_mgmt.models import (
-    App,
-    Channel,
-    ChannelChoices,
-    ErrorLog,
-    Group,
-    GroupDataRule,
-    LoginModule,
-    Menu,
-    OperationLog,
-    Role,
-    User,
-    UserRule,
-)
-from apps.system_mgmt.models.system_settings import SystemSettings
-from apps.system_mgmt.otp_challenge import (
-    check_rate_limit,
-    create_challenge,
-    invalidate_challenge,
-    record_failed_attempt,
-    reset_rate_limit,
-    verify_challenge,
-)
-from apps.system_mgmt.services.role_manage import RoleManage
-from apps.system_mgmt.utils.bk_user_utils import get_bk_user_info
-from apps.system_mgmt.utils.channel_utils import (
-    send_by_custom_webhook,
-    send_by_dingtalk_bot,
-    send_by_feishu_bot,
-    send_by_wecom_bot,
-    send_email,
-    send_email_to_user,
-    send_nats_message,
-)
-from apps.system_mgmt.utils.group_utils import GroupUtils
-from apps.system_mgmt.utils.password_validator import PasswordValidator
-from apps.system_mgmt.utils.pwd_policy_cache import get_pwd_policy_settings as _get_pwd_policy_settings
-from apps.system_mgmt.utils.token_blacklist import blacklist_token, is_blacklisted
+from apps.system_mgmt.nats import *  # noqa: F401,F403
+from apps.system_mgmt.nats import channels as _channels
+from apps.system_mgmt.nats import login as _login
 
 
-def _collect_ancestor_group_ids(seed_ids):
-    """
-    从 seed_ids 出发，沿 parent_id 链向上收集所有祖先组 ID（含自身）。
+def _sync_compat_globals():
+    _channels.send_nats_message = send_nats_message
+    _channels._normalize_nats_content = _normalize_nats_content
+    _login._get_pwd_policy_settings = _get_pwd_policy_settings
 
-    仅使用轻量级 values_list 查询（不加载角色关联），避免全表 prefetch。
-    返回的集合可用于后续有针对性地加载完整 Group 对象。
 
-    :param seed_ids: 起始组 ID 列表（通常为 user.group_list）
-    :return: 包含 seed_ids 及其所有祖先的 ID set
-    """
-    if not seed_ids:
-        return set()
-    # 一次查询获取全表的 (id, parent_id, allow_inherit_roles)，仅传输轻量列
-    all_meta = {
-        row[0]: (row[1], row[2])
-        for row in Group.objects.values_list("id", "parent_id", "allow_inherit_roles")
-    }
-    result = set()
-    stack = list(seed_ids)
-    while stack:
-        gid = stack.pop()
-        if gid in result:
-            continue
-        result.add(gid)
-        meta = all_meta.get(gid)
-        if meta:
-            parent_id, allow_inherit = meta
-            if parent_id and parent_id not in result:
-                stack.append(parent_id)
-    return result
+def send_msg_with_channel(*args, **kwargs):
+    _sync_compat_globals()
+    return _channels.send_msg_with_channel(*args, **kwargs)
 
 
-def get_user_all_roles(user):
-    """
-    获取用户的所有角色（个人角色 + 组角色，含完整继承链）
-
-    继承规则：沿 parent_id 链向上追溯，只要父级 allow_inherit_roles=True，
-    就收集该父级的角色并继续向上，直到某层 allow_inherit_roles=False 或到达根节点为止。
-
-    :param user: User实例
-    :return: 包含所有角色ID的列表
-    """
-    # 用户直接授权的角色
-    personal_role_ids = set(user.role_list)
-
-    group_role_ids = set()
-    if user.group_list:
-        # 先收集祖先组 ID（轻量查询），再按需加载完整对象（含角色关联）
-        ancestor_ids = _collect_ancestor_group_ids(user.group_list)
-        all_groups = {
-            g.id: g
-            for g in Group.objects.prefetch_related("roles").filter(id__in=ancestor_ids)
-        }
-
-        visited = set()
-
-        def collect_roles(group_id):
-            """收集 group_id 自身角色，并沿 parent_id 链递归收集继承角色"""
-            if group_id in visited:
-                return
-            visited.add(group_id)
-
-            group = all_groups.get(group_id)
-            if not group:
-                return
-
-            # 收集自身角色
-            for role in group.roles.all():
-                group_role_ids.add(role.id)
-
-            # 向上追溯：父级 allow_inherit_roles=True 才继续继承
-            parent_id = group.parent_id
-            if parent_id:
-                parent = all_groups.get(parent_id)
-                if parent and parent.allow_inherit_roles:
-                    collect_roles(parent_id)
-
-        for gid in user.group_list:
-            collect_roles(gid)
-
-    # 合并去重
-    return list(personal_role_ids | group_role_ids)
-
-
-def _get_login_expired_seconds():
-    """获取 login_expired_time 配置值（秒）。"""
-    login_expired_time_set = SystemSettings.objects.filter(key="login_expired_time").first()
-    if login_expired_time_set:
-        return int(float(login_expired_time_set.value) * 3600)
-    return 3600 * 24
-
-
-def _build_jwt_payload(user_id):
-    """构建包含 jti 和 exp 的 JWT payload。"""
-    now = int(time.time())
-    expired_seconds = _get_login_expired_seconds()
-    return {
-        "user_id": user_id,
-        "login_time": now,
-        "jti": uuid4().hex,
-        "exp": now + expired_seconds,
-    }
-
-
-def _verify_token(token):
-    token = token.split("Basic ")[-1]
-    secret_key = os.getenv("SECRET_KEY")
-    algorithm = os.getenv("JWT_ALGORITHM", "HS256")
-    user_info = jwt.decode(token, key=secret_key, algorithms=[algorithm], options={"verify_exp": False})
-    time_now = int(time.time())
-
-    # New-format token (with jti + exp): use PyJWT exp validation + blacklist check
-    if "jti" in user_info and "exp" in user_info:
-        # Re-decode with exp verification enabled
-        jwt.decode(token, key=secret_key, algorithms=[algorithm])
-        if is_blacklisted(user_info["jti"]):
-            raise Exception("Token has been revoked")
-    else:
-        # Legacy token: fall back to login_time-based expiry check
-        login_expired_time_set = SystemSettings.objects.filter(key="login_expired_time").first()
-        login_expired_time = 3600 * 24
-        if login_expired_time_set:
-            login_expired_time = float(login_expired_time_set.value) * 3600
-        if time_now - login_expired_time > user_info["login_time"]:
-            raise Exception("Token is invalid")
-
-    user = User.objects.filter(id=user_info["user_id"]).first()
-    if not user:
-        raise Exception(VERIFY_TOKEN_USER_NOT_FOUND_MESSAGE)
-    return user
-
-
-@nats_client.register
-def get_pilot_permission_by_token(token, bot_id, group_list):
-    try:
-        user = _verify_token(token)
-    except Exception:
-        return {"result": False}
-
-    # 获取用户所有角色（个人角色 + 组角色）
-    all_role_ids = get_user_all_roles(user)
-    role_list = Role.objects.filter(id__in=all_role_ids)
-    role_names = {f"{role.app}--{role.name}" if role.app else role.name for role in role_list}
-    if {"admin", "system-manager--admin", "opspilot--admin"}.intersection(role_names):
-        return {"result": True, "data": {"username": user.username}}
-    real_groups = set(group_list).intersection(user.group_list)
-    if not real_groups:
-        return {"result": False}
-    rules = UserRule.objects.filter(
-        username=user.username,
-        domain=user.domain,
-        group_rule__app="opspilot",
-        group_rule__group_id__in=list(real_groups),
-    )
-    if not rules:
-        return {"result": True, "data": {"username": user.username}}
-    for i in rules:
-        rule_obj = i.group_rule.rules.get("bot")
-        if rule_obj is None:
-            return {"result": True, "data": {"username": user.username}}
-        bot_ids = [u["id"] for u in rule_obj]
-        if bot_id in bot_ids or 0 in bot_ids:
-            return {"result": True, "data": {"username": user.username}}
-    return {"result": False}
-
-
-@nats_client.register
-def verify_token(token):
-    if not token:
-        return {"result": False, "message": "Token is missing"}
-
-    try:
-        user = _verify_token(token)
-    except Exception as e:
-        error_message = str(e)
-        return_data = {"result": False, "message": error_message}
-        if error_message == VERIFY_TOKEN_USER_NOT_FOUND_MESSAGE:
-            return_data["error_code"] = VERIFY_TOKEN_USER_NOT_FOUND_CODE
-        return return_data
-
-    # 命中缓存直接返回，跳过全量数据库查询
-    cached = get_cached_token_info(user.username, user.domain)
-    if cached is not None:
-        return cached
-
-    # 获取用户所有角色（个人角色 + 组角色）
-    all_role_ids = get_user_all_roles(user)
-
-    role_list = Role.objects.filter(id__in=all_role_ids)
-    role_names = [f"{role.app}--{role.name}" if role.app else role.name for role in role_list]
-
-    is_superuser = "admin" in role_names or "system-manager--admin" in role_names
-
-    if is_superuser:
-        # 超管需要完整组树：单次全量查询（含角色关联）
-        queryset = list(Group.objects.prefetch_related("roles").all().order_by("id"))
-        groups = [{"id": g.id, "name": g.name, "parent_id": g.parent_id} for g in queryset]
-    else:
-        # 非超管：仅加载用户直属组及其祖先（供树路径展示），消除全表扫描
-        visible_ids = _collect_ancestor_group_ids(user.group_list)
-        queryset = list(
-            Group.objects.prefetch_related("roles").filter(id__in=visible_ids).order_by("id")
-        )
-        groups = [
-            {"id": g.id, "name": g.name, "parent_id": g.parent_id}
-            for g in queryset
-            if g.id in set(user.group_list)
-        ]
-
-    # 构建嵌套组结构（queryset 已按范围过滤，避免全表扫描）
-    groups_data = GroupUtils.build_group_tree(queryset, is_superuser, [i["id"] for i in groups])
-
-    menus = cache.get(f"menus-user:{user.id}")
-
-    if not menus:
-        menus = {}
-        if not is_superuser:
-            menu_list = role_list.values_list("menu_list", flat=True)
-            menu_ids = []
-            for i in menu_list:
-                menu_ids.extend(i)
-            menu_data = Menu.objects.filter(id__in=list(set(menu_ids))).values_list("app", "name")
-            for app, name in menu_data:
-                menus.setdefault(app, []).append(name)
-
-        cache.set(f"menus-user:{user.id}", menus, 60)
-
-    result = {
-        "result": True,
-        "data": {
-            "username": user.username,
-            "display_name": user.display_name,
-            "domain": user.domain,
-            "email": user.email,
-            "is_superuser": is_superuser,
-            "group_list": groups,
-            "group_tree": groups_data,
-            "roles": role_names,
-            "role_ids": all_role_ids,
-            "locale": user.locale,
-            "permission": menus,
-            "timezone": user.timezone,
-        },
-    }
-    set_cached_token_info(user.username, user.domain, result)
-    return result
-
-
-@nats_client.register
-def revoke_token(token):
-    """撤销 token：将 jti 加入黑名单并清除用户验证缓存。"""
-    if not token:
-        return {"result": False, "message": "Token is missing"}
-
-    try:
-        token = token.split("Basic ")[-1]
-        secret_key = os.getenv("SECRET_KEY")
-        algorithm = os.getenv("JWT_ALGORITHM", "HS256")
-        user_info = jwt.decode(token, key=secret_key, algorithms=[algorithm], options={"verify_exp": False})
-    except Exception as e:
-        return {"result": False, "message": f"Invalid token: {e}"}
-
-    jti = user_info.get("jti")
-    exp = user_info.get("exp")
-    if not jti or not exp:
-        return {"result": False, "message": "Token does not support revocation (missing jti/exp)"}
-
-    blacklist_token(jti, exp)
-
-    # Clear verify_token cache for this user
-    user = User.objects.filter(id=user_info.get("user_id")).first()
-    if user:
-        clear_token_info_cache(user.username, user.domain)
-
-    return {"result": True, "message": "Token revoked"}
-
-
-@nats_client.register
-def get_user_menus(client_id, roles, username, is_superuser):
-    client = RoleManage()
-    # client_id used directly below
-    menus = []
-    if not is_superuser:
-        menu_ids = []
-        role_menus = Role.objects.filter(app=client_id, id__in=roles).values_list("menu_list", flat=True)
-        for i in role_menus:
-            menu_ids.extend(i)
-        menus = list(Menu.objects.filter(app=client_id, id__in=list(set(menu_ids))).values_list("name", flat=True))
-    user_menus = client.get_all_menus(client_id, user_menus=menus, username=username, is_superuser=is_superuser)
-    return {"result": True, "data": user_menus}
-
-
-@nats_client.register
-def get_client(client_id="", username="", domain="domain.com"):
-    app_list = App.objects.all()
-
-    if client_id:
-        app_list = app_list.filter(name__in=client_id.split(";"))
-
-    if username:
-        user = User.objects.filter(username=username, domain=domain).first()
-        if not user:
-            return {"result": False, "message": "User not found"}
-        # 获取用户所有角色（个人角色 + 组角色）
-        all_role_ids = get_user_all_roles(user)
-        role_list = Role.objects.filter(id__in=all_role_ids)
-        role_names = [f"{role.app}--{role.name}" if role.app else role.name for role in role_list]
-        is_superuser = "admin" in role_names or "system-manager--admin" in role_names
-        if not is_superuser:
-            app_name_list = list(Role.objects.filter(id__in=all_role_ids).values_list("app", flat=True).distinct())
-            if "" not in app_name_list:
-                app_list = app_list.filter(name__in=app_name_list)
-    return_data = list(app_list.order_by("id").values())
-    return {"result": True, "data": return_data}
-
-
-@nats_client.register
-def get_client_detail(client_id):
-    app_obj = App.objects.filter(name=client_id).first()
-    if not app_obj:
-        return {"result": False, "message": "Client not found"}
-    return {
-        "result": True,
-        "data": {
-            "id": app_obj.id,
-            "name": app_obj.name,
-            "display_name": app_obj.display_name,
-            "description": app_obj.description,
-            "description_cn": app_obj.description_cn,
-        },
-    }
-
-
-@nats_client.register
-def get_group_users(group=None, include_children=False):
-    """
-    获取组织下的用户列表
-    :param group: 组织ID，如果为None则返回所有用户
-    :param include_children: 是否包含子组织的用户
-    :return: 用户列表
-    """
-    if not group:
-        # 如果没有指定组织，返回所有用户
-        users = User.objects.all().values("id", "username", "display_name")
-    elif include_children:
-        group_ids = GroupUtils.get_group_with_descendants(group)
-        users = User.objects.filter(group_list__overlap=group_ids).values("id", "username", "display_name")
-    else:
-        users = User.objects.filter(group_list__contains=int(group)).values("id", "username", "display_name")
-    return {"result": True, "data": list(users)}
-
-
-def _get_actor_user_scope(actor_context, include_children=False):
-    """
-    根据调用方上下文解析允许访问的组织范围。
-
-    :param actor_context: 调用方上下文，包含 username、domain、current_team、is_superuser 等字段
-    :param include_children: 是否包含当前组织下的已授权子组织
-    :return: (user_obj, authorized_groups)
-        - user_obj: 当前调用用户对象，不存在时返回 None
-        - authorized_groups: 当前调用方允许访问的组织 ID 列表
-    """
-    username = (actor_context or {}).get("username")
-    domain = (actor_context or {}).get("domain", "domain.com")
-    current_team = (actor_context or {}).get("current_team")
-    is_superuser = (actor_context or {}).get("is_superuser", False)
-    actor_group_list = (actor_context or {}).get("group_list")
-
-    if not username or current_team in (None, ""):
-        return None, []
-
-    user_obj = User.objects.filter(username=username, domain=domain).first()
-    if not user_obj:
-        return None, []
-
-    try:
-        current_team = int(current_team)
-    except (TypeError, ValueError):
-        return user_obj, []
-
-    if is_superuser:
-        if include_children:
-            return user_obj, GroupUtils.get_group_with_descendants(current_team)
-        return user_obj, [current_team]
-
-    user_group_list = actor_group_list if actor_group_list else user_obj.group_list
-    authorized_groups = GroupUtils.get_user_authorized_child_groups(
-        user_group_list,
-        current_team,
-        include_children=include_children,
-    )
-    return user_obj, authorized_groups
-
-
-@nats_client.register
-def get_group_users_scoped(actor_context, group=None, include_children=False):
-    """
-    在调用方授权范围内查询组织用户列表。
-
-    :param actor_context: 调用方上下文，包含 username、domain、current_team、is_superuser 等字段
-    :param group: 可选，指定查询的组织 ID；若不传则使用调用方当前授权范围
-    :param include_children: 是否包含目标组织下的已授权子组织用户
-    :return: 标准 NATS 返回结构，data 为用户列表
-    """
-    user_obj, authorized_groups = _get_actor_user_scope(actor_context, include_children=include_children)
-    if not user_obj or not authorized_groups:
-        return {"result": True, "data": []}
-
-    if group is not None:
-        try:
-            group = int(group)
-        except (TypeError, ValueError):
-            return {"result": True, "data": []}
-        if group not in authorized_groups:
-            return {"result": True, "data": []}
-        query_groups = GroupUtils.get_user_authorized_child_groups(
-            user_obj.group_list,
-            group,
-            include_children=include_children,
-        )
-    else:
-        query_groups = authorized_groups
-
-    if not query_groups:
-        return {"result": True, "data": []}
-
-    user_filter = Q()
-    for group_id in query_groups:
-        user_filter |= Q(group_list__contains=int(group_id))
-    users = User.objects.filter(user_filter).values("id", "username", "display_name")
-    return {"result": True, "data": list(users)}
-
-
-@nats_client.register
-def get_authorized_groups_scoped(actor_context, include_children=False):
-    """返回调用方在当前组织上下文下可访问的组织范围。"""
-    _user_obj, authorized_groups = _get_actor_user_scope(actor_context, include_children=include_children)
-    return {"result": True, "data": authorized_groups}
-
-
-@nats_client.register
-def get_all_users():
-    data = User.objects.all().values(*User.display_fields())
-    return {"result": True, "data": list(data)}
-
-
-@nats_client.register
-def search_groups(query_params):
-    groups = Group.objects.filter(name__contains=query_params["search"]).values()
-    return {"result": True, "data": list(groups)}
-
-
-@nats_client.register
-def search_users(query_params):
-    page = int(query_params.get("page", 1))
-    page_size = int(query_params.get("page_size", 10))
-    search = query_params.get("search", "")
-    queryset = User.objects.filter(Q(username__icontains=search) | Q(display_name__icontains=search) | Q(email__icontains=search))
-    start = (page - 1) * page_size
-    end = page * page_size
-    total = queryset.count()
-    display_fields = User.display_fields() + ["group_list"]
-    data = queryset.values(*display_fields)[start:end]
-    return {"result": True, "data": {"count": total, "users": list(data)}}
-
-
-@nats_client.register
-def init_user_default_attributes(user_id, group_name, default_group_id):
-    try:
-        role_ids = list(
-            Role.objects.filter(name="guest", app__in=["opspilot", "cmdb", "monitor", "alarm", "log", "node", "mlops", "job"]).values_list(
-                "id", flat=True
-            )
-        )
-        normal_role = Role.objects.get(name="normal", app="opspilot")
-        user = User.objects.get(id=user_id)
-        top_group, _ = Group.objects.get_or_create(
-            name=os.getenv("DEFAULT_GROUP_NAME", "Guest"),
-            parent_id=0,
-            defaults={"description": ""},
-        )
-        if Group.objects.filter(parent_id=top_group.id, name=group_name).exists():
-            return {"result": False, "message": "Group already exists"}
-
-        guest_group, _ = Group.objects.get_or_create(name="OpsPilotGuest", parent_id=0)
-        with transaction.atomic():
-            group_obj = Group.objects.create(name=group_name, parent_id=top_group.id)
-            user.locale = "zh-Hans"
-            user.timezone = "Asia/Shanghai"
-            user.role_list.extend(role_ids)
-            user.role_list = list(set(user.role_list))  # 去重
-            if normal_role.id in user.role_list:
-                user.role_list.remove(normal_role.id)
-            user.group_list.remove(int(default_group_id))
-            user.group_list.append(guest_group.id)
-            user.group_list.append(group_obj.id)
-            user.save()
-            set_opspilot_guest_group_default_rule(guest_group, user)
-        cache.delete(f"group_{user.username}")
-        return {"result": True, "data": {"group_id": group_obj.id}}
-    except Exception as e:
-        logger.exception(e)
-        return {"result": False, "message": str(e)}
-
-
-@nats_client.register
-def create_guest_role():
-    app_map = {
-        "opspilot": OPSPILOT_GUEST_MENUS[:],
-        "cmdb": CMDB_MENUS[:],
-        "monitor": MONITOR_MENUS[:],
-    }
-    guest_group, _ = Group.objects.get_or_create(name="Guest", parent_id=0, defaults={"description": "Guest group"})
-    app_guest_group, _ = Group.objects.get_or_create(name="OpsPilotGuest", parent_id=0)
-    for app, app_menus in app_map.items():
-        menus = dict(Menu.objects.filter(app=app).values_list("id", "name"))
-        menu_list = [k for k, v in menus.items() if v in app_menus]
-        Role.objects.update_or_create(name="guest", app=app, defaults={"menu_list": menu_list})
-    return {"result": True, "data": {"group_id": app_guest_group.id}}
-
-
-@nats_client.register
-def create_default_rule(llm_model, ocr_model, embed_model, rerank_model):
-    guest_group = Group.objects.get(name="OpsPilotGuest", parent_id=0)
-    GroupDataRule.objects.get_or_create(
-        name="OpsPilot内置规则",
-        app="opspilot",
-        defaults=dict(
-            group_id=guest_group.id,
-            description="Guest组数据权限规则",
-            group_name=guest_group.name,
-            rules={
-                "skill": [{"id": 0, "name": "All", "permission": ["View"]}],
-                "tools": [{"id": 0, "name": "All", "permission": ["View"]}],
-                "provider": {
-                    "llm_model": [
-                        {
-                            "id": llm_model["id"],
-                            "name": llm_model["name"],
-                            "permission": ["View"],
-                        }
-                    ],
-                    "ocr_model": [{"id": i["id"], "name": i["name"], "permission": ["View"]} for i in ocr_model],
-                    "embed_model": [{"id": i["id"], "name": i["name"], "permission": ["View"]} for i in embed_model],
-                    "rerank_model": [
-                        {
-                            "id": rerank_model["id"],
-                            "name": rerank_model["name"],
-                            "permission": ["View"],
-                        }
-                    ],
-                },
-                "knowledge": [{"id": 0, "name": "All", "permission": ["View"]}],
-            },
-        ),
-    )
-    return {"result": True}
-
-
-@nats_client.register
-def get_all_groups():
-    groups = Group.objects.prefetch_related("roles").all()
-    return_data = GroupUtils.build_group_tree(groups, True)
-    return {"result": True, "data": return_data}
-
-
-@nats_client.register
-def get_channel_detail(channel_id):
-    channel_obj = Channel.objects.filter(id=channel_id).first()
-    if not channel_obj:
-        return {"result": False, "message": "传入的channel_id无法匹配到channel"}
-    return_data = {
-        "name": channel_obj.name,
-        "description": channel_obj.description,
-        "config": channel_obj.config,
-        "team": channel_obj.team,
-        "channel_type": channel_obj.channel_type,
-    }
-    return {"result": True, "data": return_data}
-
-
-@nats_client.register
-def search_channel_list(channel_type="", teams=None, include_children=False):
-    """
-    :param channel_type: str， 目前只有email、enterprise_wechat_bot
-    :param teams: list, [1,2,3]
-    :param include_children: bool , True、False
-    """
-    # 空 teams 直接返回空数据
-    if not teams:
-        return {"result": True, "data": []}
-
-    # 如果 include_children 为 True，递归获取所有子组织
-    if include_children:
-        # 一次性获取所有组织，避免递归查询数据库
-        all_groups = Group.objects.values_list("id", "parent_id")
-        # 构建 parent_id -> [child_ids] 的映射
-        children_map = {}
-        for gid, pid in all_groups:
-            if pid is not None:
-                children_map.setdefault(pid, []).append(gid)
-
-        # 在内存中递归获取所有子组织
-        def get_descendants(group_id, result_set):
-            result_set.add(group_id)
-            for child_id in children_map.get(group_id, []):
-                get_descendants(child_id, result_set)
-
-        all_teams = set()
-        for team_id in teams:
-            get_descendants(team_id, all_teams)
-        teams = list(all_teams)
-
-    # 构建 teams 筛选条件：team 字段与 teams 有交集
-    channels = Channel.objects.all()
-    if channel_type:
-        channels = channels.filter(channel_type=channel_type)
-
-    # 使用 Q 对象构建 OR 条件
-    if teams:
-        team_filter = Q(team__contains=teams[0])
-        for team_id in teams[1:]:
-            team_filter |= Q(team__contains=team_id)
-        channels = channels.filter(team_filter)
-
-    return {
-        "result": True,
-        "data": [i for i in channels.values("id", "name", "channel_type", "description")],
-    }
-
-
-@nats_client.register
-def search_channel_list_scoped(actor_context, channel_type="", teams=None, include_children=False):
-    """
-    在调用方授权范围内查询通知通道列表。
-
-    :param actor_context: 调用方上下文，包含 username、domain、current_team、is_superuser 等字段
-    :param channel_type: 可选，通道类型过滤条件
-    :param teams: 可选，待查询的组织 ID 列表；最终会与调用方授权范围取交集
-    :param include_children: 是否包含当前组织下的已授权子组织
-    :return: 标准 NATS 返回结构，data 为通知通道列表
-    """
-    user_obj, authorized_groups = _get_actor_user_scope(actor_context, include_children=include_children)
-    if not user_obj or not authorized_groups:
-        return {"result": True, "data": []}
-
-    if teams:
-        normalized_teams = []
-        for team in teams:
-            try:
-                normalized_teams.append(int(team))
-            except (TypeError, ValueError):
-                continue
-        teams = [team for team in normalized_teams if team in authorized_groups]
-    else:
-        teams = authorized_groups
-
-    return search_channel_list(
-        channel_type=channel_type,
-        teams=teams,
-        include_children=False,
-    )
-
-
-def _resolve_message_receivers(receivers):
-    if not receivers:
-        return None
-
-    if all(isinstance(r, int) or (isinstance(r, str) and r.isdigit()) for r in receivers):
-        return User.objects.filter(id__in=[int(r) for r in receivers])
-
-    if all(isinstance(r, str) and r.strip() and not r.isdigit() for r in receivers):
-        return User.objects.filter(username__in=[receiver.strip() for receiver in receivers])
-
-    return None
-
-
-def _normalize_nats_content(content):
-    if isinstance(content, str):
-        try:
-            content = json.loads(content)
-        except (json.JSONDecodeError, TypeError):
-            return None, {"result": False, "message": "NATS content is not valid JSON"}
-
-    if not isinstance(content, dict):
-        return None, {"result": False, "message": "NATS content must be a dict"}
-
-    message = content.get("message")
-    if not isinstance(message, str) or not message.strip():
-        return None, {"result": False, "message": "NATS content.message must be a non-empty string"}
-
-    team = content.get("team")
-    # team 现在只允许单个组织 ID；兼容历史的单元素列表写法
-    if isinstance(team, (list, tuple)):
-        if len(team) != 1:
-            return None, {"result": False, "message": "NATS content.team must be a single team id"}
-        team = team[0]
-    team_value = str(team).strip()
-    if not team_value or not team_value.isdigit():
-        return None, {"result": False, "message": "NATS content.team must be a single integer team id"}
-    normalized_team = int(team_value)
-
-    user_ids = content.get("user_ids")
-    if not isinstance(user_ids, list):
-        return None, {"result": False, "message": "NATS content.user_ids must be a list"}
-
-    normalized_user_ids = []
-    for user_id in user_ids:
-        if user_id is None:
-            continue
-        normalized_user_id = str(user_id).strip()
-        if normalized_user_id:
-            normalized_user_ids.append(normalized_user_id)
-
-    return {
-        "message": message.strip(),
-        "team": normalized_team,
-        "user_ids": normalized_user_ids,
-    }, None
-
-
-# NATS 通道里「原样透传 content」的内部方法白名单。
-# 这些方法（如告警中心 receive_alert_events）有自己的入参契约（如 {source_id, pusher, events}），
-# 不走 OpsPilot/IM 触发那套 message/team/user_ids 规范化，否则 content 会被裁剪、事件丢失。
-RAW_PASSTHROUGH_NATS_METHODS = {"receive_alert_events"}
-
-
-@nats_client.register
-def send_msg_with_channel(channel_id, title, content, receivers, attachments=None):
-    """
-    通过指定通道发送消息
-    :param channel_id: 通道ID
-    :param title: 邮件主题（企微机器人传空字符串即可）
-    :param content: 正文内容
-    :param receivers: 用户ID列表 [1, 2, 3, 4] 或用户名列表 ["user1", "user2"]
-    :param attachments: 附件列表（仅email通道支持），格式为:
-        [{"filename": "文件名.pdf", "content": "base64编码的文件内容"}, ...]
-        注意: 附件内容必须是base64编码的字符串，因为NATS使用JSON序列化传输
-    """
-    channel_obj = Channel.objects.filter(id=channel_id).first()
-    if not channel_obj:
-        return {"result": False, "message": "Channel not found"}
-    # 兼容用户ID列表和用户名列表两种情况
-    user_list = _resolve_message_receivers(receivers)
-    if channel_obj.channel_type == ChannelChoices.EMAIL:
-        # 邮件发送需要校验收件人是否存在
-        if not user_list or not user_list.exists():
-            return {"result": False, "message": "No valid recipients found"}
-        return send_email(channel_obj, title, content, user_list, attachments)
-    elif channel_obj.channel_type == ChannelChoices.ENTERPRISE_WECHAT_BOT:
-        if user_list is not None:
-            display_names = list(user_list.values_list("display_name", flat=True))
-        else:
-            display_names = receivers if isinstance(receivers, list) else [receivers]
-        return send_by_wecom_bot(channel_obj, content, display_names)
-    elif channel_obj.channel_type == ChannelChoices.FEISHU_BOT:
-        if user_list is not None:
-            display_names = list(user_list.values_list("display_name", flat=True))
-        else:
-            display_names = receivers if isinstance(receivers, list) else [receivers]
-        return send_by_feishu_bot(channel_obj, title, content, display_names)
-    elif channel_obj.channel_type == ChannelChoices.DINGTALK_BOT:
-        if user_list is not None:
-            display_names = list(user_list.values_list("display_name", flat=True))
-        else:
-            display_names = receivers if isinstance(receivers, list) else [receivers]
-        return send_by_dingtalk_bot(channel_obj, title, content, display_names)
-    elif channel_obj.channel_type == ChannelChoices.CUSTOM_WEBHOOK:
-        return send_by_custom_webhook(channel_obj, content, receivers)
-    elif channel_obj.channel_type == ChannelChoices.NATS:
-        # NATS 通道：content 作为 kwargs 传递给目标服务
-        method_name = (channel_obj.config or {}).get("method_name")
-        if method_name in RAW_PASSTHROUGH_NATS_METHODS:
-            # 内部直推通道（如告警中心）：原样透传 content，跳过 IM 触发的字段规范化。
-            return send_nats_message(channel_obj, content)
-        normalized, error = _normalize_nats_content(content)
-        if error:
-            return error
-        return send_nats_message(channel_obj, normalized)
-    return {"result": False, "message": "Unsupported channel type"}
-    # return send_wechat(channel_obj, content, user_list)
-
-
-# OpsPilot 工作流自动托管的 NATS 触发通道：靠 config.source 标识，禁止用户在通道管理里编辑/删除
-OPSPILOT_CHANNEL_SOURCE = "opspilot"
-# 通道路由到 OpsPilot 的 NATS namespace，需与部署的 NATS_NAMESPACE 一致（默认 bklite），
-# 而非字符串 "opspilot"——所有 @nats_client.register 方法都经该 namespace 路由。
-OPSPILOT_NATS_NAMESPACE = os.getenv("NATS_NAMESPACE", "bklite")
-OPSPILOT_NATS_METHOD = "trigger_workflow_by_nats"
-
-
-def _list_opspilot_nats_channels(bot_id):
-    """返回某个 bot 名下、由 OpsPilot 托管的 NATS 通道（DB 无关，Python 侧过滤 config）。"""
-    channels = Channel.objects.filter(channel_type=ChannelChoices.NATS)
-    result = []
-    for channel in channels:
-        config = channel.config or {}
-        if config.get("source") == OPSPILOT_CHANNEL_SOURCE and str(config.get("bot_id")) == str(bot_id):
-            result.append(channel)
-    return result
-
-
-@nats_client.register
-def sync_opspilot_nats_channels(bot_id, bot_name, team, nodes, timeout=60):
-    """对账 OpsPilot 某个 bot 的 NATS 触发节点对应的通道（增/改/删）。
-
-    :param bot_id: Bot ID
-    :param bot_name: Bot 名称（用于拼通道名）
-    :param team: 通道归属组织 ID 列表
-    :param nodes: [{"node_id": "xxx", "name": "节点label"}, ...]
-    :param timeout: NATS 请求超时（秒）
-    """
-    try:
-        bot_id = int(bot_id)
-    except (TypeError, ValueError):
-        return {"result": False, "message": "bot_id must be an integer"}
-
-    team = team or []
-    nodes = nodes or []
-    description = "OpsPilot 工作流自动创建的 NATS 触发通道"
-
-    existing_by_node = {(ch.config or {}).get("node_id"): ch for ch in _list_opspilot_nats_channels(bot_id)}
-
-    incoming_node_ids = set()
-    created = updated = 0
-    for node in nodes:
-        node_id = str((node or {}).get("node_id") or "").strip()
-        if not node_id:
-            continue
-        incoming_node_ids.add(node_id)
-        label = str((node or {}).get("name") or node_id).strip()
-        # 通道名：BOT名 - 节点名；Channel.name 上限 100
-        name = f"{bot_name} - {label}"[:100]
-        config = {
-            "namespace": OPSPILOT_NATS_NAMESPACE,
-            "method_name": OPSPILOT_NATS_METHOD,
-            "bot_id": bot_id,
-            "node_id": node_id,
-            "timeout": timeout,
-            "source": OPSPILOT_CHANNEL_SOURCE,
-        }
-        channel = existing_by_node.get(node_id)
-        if channel:
-            channel.name = name
-            channel.config = config
-            channel.team = team
-            channel.description = description
-            channel.save()
-            updated += 1
-        else:
-            Channel.objects.create(
-                name=name,
-                channel_type=ChannelChoices.NATS,
-                config=config,
-                team=team,
-                description=description,
-            )
-            created += 1
-
-    # 对账删除：flow_json 里已不存在的旧节点对应的通道
-    deleted = 0
-    for node_id, channel in existing_by_node.items():
-        if node_id not in incoming_node_ids:
-            channel.delete()
-            deleted += 1
-
-    return {"result": True, "data": {"created": created, "updated": updated, "deleted": deleted}}
-
-
-@nats_client.register
-def delete_opspilot_nats_channels(bot_id):
-    """删除某个 bot 名下所有 OpsPilot 托管的 NATS 通道（bot 删除时清理）。"""
-    try:
-        bot_id = int(bot_id)
-    except (TypeError, ValueError):
-        return {"result": False, "message": "bot_id must be an integer"}
-
-    deleted = 0
-    for channel in _list_opspilot_nats_channels(bot_id):
-        channel.delete()
-        deleted += 1
-    return {"result": True, "data": {"deleted": deleted}}
-
-
-@nats_client.register
-def search_opspilot_nats_channels(teams=None, bot_id=None, include_children=False):
-    """查询 OpsPilot 托管的 NATS 触发通道（config.source == "opspilot"）。
-
-    与通用 search_channel_list 不同：本接口专门按 OpsPilot 托管标识过滤，
-    支持跨团队/全局列举，并返回路由字段 bot_id / node_id。
-
-    :param teams: 可选，组织 ID 列表；为空/None 则跨团队全局列举
-    :param bot_id: 可选，仅返回该 Bot 的通道
-    :param include_children: 当传 teams 时，是否一并包含其子组织
-    :return: 标准 NATS 返回结构，data 为 [{id, name, description, team, bot_id, node_id}]
-    """
-    channels = Channel.objects.filter(channel_type=ChannelChoices.NATS)
-
-    # 传了 teams 才按组织过滤；为空表示全局
-    if teams:
-        normalized_teams = []
-        for team_id in teams:
-            try:
-                normalized_teams.append(int(team_id))
-            except (TypeError, ValueError):
-                continue
-
-        if include_children and normalized_teams:
-            all_groups = Group.objects.values_list("id", "parent_id")
-            children_map = {}
-            for gid, pid in all_groups:
-                if pid is not None:
-                    children_map.setdefault(pid, []).append(gid)
-
-            def _collect_descendants(group_id, acc):
-                acc.add(group_id)
-                for child_id in children_map.get(group_id, []):
-                    _collect_descendants(child_id, acc)
-
-            expanded = set()
-            for team_id in normalized_teams:
-                _collect_descendants(team_id, expanded)
-            normalized_teams = list(expanded)
-
-        if not normalized_teams:
-            return {"result": True, "data": []}
-
-        team_filter = Q(team__contains=normalized_teams[0])
-        for team_id in normalized_teams[1:]:
-            team_filter |= Q(team__contains=team_id)
-        channels = channels.filter(team_filter)
-
-    # DB 无关：在 Python 侧按 config.source（及可选 bot_id）过滤
-    data = []
-    for channel in channels:
-        config = channel.config or {}
-        if config.get("source") != OPSPILOT_CHANNEL_SOURCE:
-            continue
-        if bot_id is not None and str(config.get("bot_id")) != str(bot_id):
-            continue
-        data.append(
-            {
-                "id": channel.id,
-                "name": channel.name,
-                "description": channel.description,
-                "team": channel.team,
-                "bot_id": config.get("bot_id"),
-                "node_id": config.get("node_id"),
-            }
-        )
-    return {"result": True, "data": data}
-
-
-@nats_client.register
-def send_email_to_receiver(title, content, receiver):
-    channel_obj = Channel.objects.filter(channel_type=ChannelChoices.EMAIL).first()
-    channel_config = channel_obj.config
-    channel_obj.decrypt_field("smtp_pwd", channel_config)
-    return send_email_to_user(channel_config, content, [receiver], title)
-
-
-@nats_client.register
-def get_user_rules(group_id, username):
-    rules = UserRule.objects.filter(username=username).filter(Q(group_rule__group_id=group_id) | Q(group_rule__group_name="OpsPilotGuest"))
-    if not rules:
-        return {}
-    return_data = {}
-    for i in rules:
-        if i.group_rule.group_name == "OpsPilotGuest":
-            return_data.setdefault(i.group_rule.app, {})["guest"] = i.group_rule.rules
-        else:
-            return_data.setdefault(i.group_rule.app, {})["normal"] = i.group_rule.rules
-    return return_data
-
-
-def _prepare_user_rules_query(group_id, username, domain, app, include_children=False):
-    """
-    准备用户权限规则查询的通用逻辑
-    :param group_id: 组ID
-    :param username: 用户名
-    :param domain: 域
-    :param app: 应用名称
-    :param include_children: 是否包含子组（递归查询所有子孙组）
-    :return: (user_obj, query_group_ids, admin_teams, has_guest_group, is_admin)
-    """
-    # 获取用户对象
-    user_obj = User.objects.filter(username=username, domain=domain).first()
-    if not user_obj:
-        return None, None, None, None, None
-
-    # 获取管理员角色列表
-    admin_list = list(Role.objects.filter(name="admin").filter(Q(app="") | Q(app=app)).values_list("id", flat=True))
-
-    # 获取用户所有角色（个人角色 + 组角色）
-    all_role_ids = get_user_all_roles(user_obj)
-    is_admin = bool(set(all_role_ids).intersection(admin_list))
-
-    # 获取查询的组ID列表（包含子组）
-    query_group_ids = []
-    if include_children:
-        # 使用优化后的单次查询方法替代 N+1 的 get_all_child_groups
-        query_group_ids = GroupUtils.get_group_with_descendants_filtered(int(group_id), group_list=user_obj.group_list)
-
-    query_group_ids.append(int(group_id))
-    query_group_ids = list(set(query_group_ids))
-    # 设置管理员团队
-    admin_teams = query_group_ids[:]
-
-    # 检查是否有guest组权限
-    guest_group = Group.objects.filter(name="OpsPilotGuest").first()
-    has_guest_group = False
-    if guest_group and guest_group.id in user_obj.group_list:
-        has_guest_group = True
-        admin_teams.append(guest_group.id)
-
-    return user_obj, query_group_ids, admin_teams, has_guest_group, is_admin
-
-
-@nats_client.register
-def get_user_rules_by_module(group_id, username, domain, app, module, include_children=False):
-    """
-    获取用户在指定模块下的所有权限规则，按子模块分组返回
-    :param group_id: 组ID
-    :param username: 用户名
-    :param domain: 域
-    :param app: 应用名称
-    :param module: 模块名称
-    :param include_children: 是否包含子组（递归查询所有子孙组）
-    """
-    # 使用通用查询准备函数
-    user_obj, query_group_ids, admin_teams, has_guest_group, is_admin = _prepare_user_rules_query(group_id, username, domain, app, include_children)
-
-    if not user_obj:
-        return {"result": False, "message": "User not found"}
-
-    all_permission = {"all": {"instance": [], "team": admin_teams}}
-
-    # 如果是管理员，返回所有权限
-    if is_admin:
-        return {"result": True, "data": all_permission, "team": admin_teams}
-
-    # 构建查询过滤条件
-    if has_guest_group:
-        base_filter = Q(group_rule__group_id__in=query_group_ids) | Q(group_rule__group_name="OpsPilotGuest")
-    else:
-        base_filter = Q(group_rule__group_id__in=query_group_ids)
-    module_filter = Q(group_rule__rules__has_key=module)
-
-    rules = UserRule.objects.filter(username=username, domain=domain, group_rule__app=app).filter(base_filter & module_filter)
-    if not rules:
-        # 普通用户无权限规则时，返回空数据，但保留 team 用于组织过滤
-        # 注意：不返回 all_permission，避免被误认为管理员权限
-        return {"result": True, "data": {}, "team": admin_teams}
-
-    result = {}
-    group_list = {i.group_rule.group_id for i in rules}
-    all_permission_team = [i for i in admin_teams if i not in group_list]
-
-    for rule in rules:
-        # 获取模块数据
-        module_data = rule.group_rule.rules.get(module, {})
-
-        # 遍历模块下的所有分类和子模块
-        for category, sub_modules in module_data.items():
-            if isinstance(sub_modules, dict):
-                # 嵌套结构（如 provider.llm_model）
-                for sub_module_id, rule_data in sub_modules.items():
-                    _accumulate_rule_result(result, sub_module_id, rule_data, rule.group_rule.group_id, all_permission_team)
-            else:
-                # 扁平结构（如 skill、bot）
-                _accumulate_rule_result(result, category, sub_modules, rule.group_rule.group_id, all_permission_team)
-
-    return {"result": True, "data": result, "team": admin_teams}
-
-
-@nats_client.register
-def get_user_rules_by_app(group_id, username, domain, app, module, child_module="", include_children=False):
-    """
-    获取用户在指定应用模块下的权限规则
-    :param group_id: 组ID
-    :param username: 用户名
-    :param domain: 域
-    :param app: 应用名称
-    :param module: 模块名称
-    :param child_module: 子模块名称
-    :param include_children: 是否包含子组（递归查询所有子孙组）
-    """
-    # 使用通用查询准备函数
-    user_obj, query_group_ids, admin_teams, has_guest_group, is_admin = _prepare_user_rules_query(group_id, username, domain, app, include_children)
-
-    if not user_obj:
-        return {"instance": [], "team": []}
-
-    # 如果是管理员，返回所有权限
-    if is_admin:
-        return {"instance": [], "team": admin_teams}
-
-    # 构建查询过滤条件
-    if has_guest_group:
-        base_filter = Q(group_rule__group_id__in=query_group_ids) | Q(group_rule__group_name="OpsPilotGuest")
-    else:
-        base_filter = Q(group_rule__group_id__in=query_group_ids)
-    # 添加模块过滤条件
-    module_filter = Q(group_rule__rules__has_key=module)
-
-    # 如果指定了子模块，不在数据库层面过滤，在Python层面处理复杂嵌套
-    rules = list(UserRule.objects.filter(username=username, domain=domain, group_rule__app=app).filter(base_filter & module_filter))
-
-    if not rules:
-        return {"instance": [], "team": admin_teams}
-
-    group_list = {i.group_rule.group_id for i in rules}
-    return_data = {
-        "instance": [],
-        "team": [i for i in admin_teams if i not in group_list],
-    }
-
-    for rule in rules:
-        # 获取模块数据
-        module_data = rule.group_rule.rules.get(module, [])
-
-        # 如果指定了子模块，获取子模块数据
-        if child_module:
-            target_data = find_child_module_data(module_data, child_module)
-        else:
-            target_data = module_data
-        # 处理规则数据
-        has_all_permission, instance_data = process_rule_data(target_data)
-
-        if has_all_permission:
-            return_data["team"].append(rule.group_rule.group_id)
-        else:
-            return_data["instance"].extend(instance_data)
-
-    return return_data
-
-
-def find_child_module_data(module_data, target_child_module):
-    """在模块数据中查找子模块数据，支持嵌套结构"""
-    if not isinstance(module_data, dict):
-        return []
-
-    # 直接查找子模块
-    if target_child_module in module_data:
-        return module_data[target_child_module]
-
-    # 在嵌套结构中查找子模块
-    for key, value in module_data.items():
-        if isinstance(value, dict) and target_child_module in value:
-            return value[target_child_module]
-    return []
-
-
-def process_rule_data(rule_data):
-    """处理规则数据，返回是否为全部权限和具体实例数据"""
-    if not rule_data:
-        return True, []
-    if isinstance(rule_data, list):
-        rule_data = [item for item in rule_data if isinstance(item, dict) and item.get("id") not in ["-1", -1]]
-        ids = [item.get("id") for item in rule_data]
-        has_all_permission = 0 in ids or "0" in ids
-        return has_all_permission, rule_data if not has_all_permission else []
-
-    return True, []
-
-
-def _accumulate_rule_result(result, key, rule_data, group_id, all_permission_team):
-    """
-    累积规则结果到指定的 key 中
-    :param result: 结果字典
-    :param key: 子模块 ID 或分类名称
-    :param rule_data: 规则数据
-    :param group_id: 组 ID
-    :param all_permission_team: 全权限团队列表
-    """
-    # 初始化结果键
-    if key not in result:
-        result[key] = {"instance": [], "team": all_permission_team[:]}
-
-    # 处理规则数据
-    has_all_permission, instance_data = process_rule_data(rule_data)
-
-    if has_all_permission:
-        # 如果有全部权限，添加组 ID 到团队列表
-        if group_id not in result[key]["team"]:
-            result[key]["team"].append(group_id)
-    else:
-        # 否则添加实例数据
-        result[key]["instance"].extend(instance_data)
-
-
-@nats_client.register
-def get_group_id(group_name):
-    group = Group.objects.filter(name=group_name, parent_id=0).first()
-    if not group:
-        return {"result": False, "message": f"group named '{group_name}' not exists."}
-    return {"result": True, "data": group.id}
-
-
-@nats_client.register
-def login(username, password):
-    user = User.objects.filter(username=username, domain="domain.com").first()
-    if not user:
-        return {"result": False, "message": "Username or password is incorrect"}
-
-    # 初始化语言加载器，使用用户的locale
-    loader = LanguageLoader(app="system_mgmt", default_lang=user.locale or "en")
-
-    # 检查账号是否被锁定
-    now = timezone.now()
-    if user.account_locked_until and user.account_locked_until > now:
-        # 计算剩余锁定时间（分钟）
-        remaining_minutes = int((user.account_locked_until - now).total_seconds() / 60) + 1
-        msg = loader.get(
-            "login.account_locked",
-            "Account is locked. Please try again after {minutes} minutes.",
-        ).format(minutes=remaining_minutes)
-        return {"result": False, "message": msg}
-
-    # 使用 check_password 验证密码是否匹配
-    if not check_password(password, user.password):
-        # 密码错误，递增错误次数
-        user.password_error_count += 1
-
-        # 批量读取密码策略（单次 DB 查询 + 缓存，避免每次失败登录多次打 DB）
-        pwd_policy = _get_pwd_policy_settings()
-        max_retry_count = pwd_policy["pwd_set_max_retry_count"]
-        lock_duration_seconds = pwd_policy["pwd_set_lock_duration"]
-
-        # 如果错误次数达到或超过最大重试次数，锁定账号
-        if user.password_error_count >= max_retry_count:
-            user.account_locked_until = now + timedelta(seconds=lock_duration_seconds)
-            user.save()
-            lock_duration_minutes = int(lock_duration_seconds / 60) + 1
-            return {
-                "result": False,
-                "message": loader.get(
-                    "login.account_locked_too_many_attempts",
-                    "Account locked due to too many failed attempts. Please try again after {minutes} minutes.",
-                ).format(minutes=lock_duration_minutes),
-            }
-
-        user.save()
-        remaining_attempts = max_retry_count - user.password_error_count
-        return {
-            "result": False,
-            "message": loader.get(
-                "login.incorrect_password_with_attempts",
-                "Username or password is incorrect. {attempts} attempts remaining.",
-            ).format(attempts=remaining_attempts),
-        }
-
-    # 密码正确，重置错误次数和锁定状态
-    user.password_error_count = 0
-    user.account_locked_until = None
-    user.save()
-
-    is_admin_account = user.username == "admin"
-
-    # 检查密码过期
-    password_expiry_reminder = ""
-    if user.password_last_modified:
-        # 批量读取密码策略（与密码错误路径共用同一缓存，无额外 DB 开销）
-        pwd_policy = _get_pwd_policy_settings()
-        validity_period_days = pwd_policy["pwd_set_validity_period"]
-
-        # validity_period_days <= 0 表示永不过期，跳过过期检查
-        if validity_period_days > 0:
-            reminder_days = pwd_policy["pwd_set_expiry_reminder_days"]
-
-            password_expire_date = user.password_last_modified + timedelta(days=validity_period_days)
-            days_until_expire = (password_expire_date - now).days
-
-            # 密码已过期，阻止登录
-            if days_until_expire <= 0:
-                if is_admin_account:
-                    if not user.temporary_pwd:
-                        user.temporary_pwd = True
-                        user.save()
-                else:
-                    return {
-                        "result": False,
-                        "message": loader.get(
-                            "login.password_expired_contact_admin",
-                            "Your password has expired. Please contact the administrator to reset your password.",
-                        ),
-                    }
-
-            # 密码快过期，生成提醒消息
-            if days_until_expire <= reminder_days:
-                password_expiry_reminder = loader.get(
-                    "login.password_expiring_soon",
-                    "Your password will expire in {days} day(s). Please change it soon.",
-                ).format(days=days_until_expire)
-
-    result = get_user_login_token(user, username, skip_token_for_otp=True)
-    if result.get("result"):
-        # Add password_expiry_reminder to response
-        result["data"]["password_expiry_reminder"] = password_expiry_reminder
-    return result
-
-
-@nats_client.register
-def reset_pwd(username, domain, password, caller_token=""):
-    """
-    重置用户密码（NATS接口）
-
-    会进行密码复杂度校验，以及调用方身份校验——caller_token 必须是有效的
-    JWT 会话 token，且 token 所属用户必须与 username 一致（自助改密），
-    以阻止任意内网服务通过 NATS 总线篡改他人密码。
-    """
-    # 调用方身份校验：验证 caller_token 有效，且属于请求的同一用户
-    if not caller_token:
-        return {"result": False, "message": "caller_token is required"}
-    try:
-        caller = _verify_token(caller_token)
-    except Exception:
-        return {"result": False, "message": "Unauthorized: invalid token"}
-    caller_domain = getattr(caller, "domain", "") or "domain.com"
-    target_domain = domain or "domain.com"
-    if caller.username != username or caller_domain != target_domain:
-        return {"result": False, "message": "Unauthorized: caller does not match target user"}
-
-    filter_kwargs = {"username": username}
-    if domain:
-        filter_kwargs["domain"] = domain
-    user = User.objects.filter(**filter_kwargs).first()
-    if not user:
-        return {"result": False, "message": "Username not exists"}
-
-    # 校验密码复杂度
-    is_valid, error_message = PasswordValidator.validate_password(password)
-    if not is_valid:
-        return {"result": False, "message": error_message}
-
-    user.password = make_password(password)
-    user.temporary_pwd = False
-    user.save()
-    return {"result": True}
-
-
-@nats_client.register
-def wechat_user_register(user_id, nick_name):
-    with transaction.atomic():
-        user, is_first_login = User.objects.select_for_update().get_or_create(
-            username=user_id, defaults={"display_name": nick_name}
-        )
-        default_group = Group.objects.filter(name="OpsPilotGuest", parent_id=0).first()
-        if not user.group_list and default_group:
-            user.group_list = [default_group.id]
-        default_role = list(
-            Role.objects.filter(
-                Q(name="normal", app__in=["opspilot", "ops-console"])
-                | Q(
-                    name="guest",
-                    app__in=["opspilot", "cmdb", "monitor", "log", "alarm", "node", "mlops", "job"],
-                )
-            ).values_list("id", flat=True)
-        )
-        default_role.extend(user.role_list)
-        user.role_list = list(set(default_role))
-        user.last_login = timezone.now()
-        user.save()
-    try:
-        if default_group:
-            set_opspilot_guest_group_default_rule(default_group, user)
-    except Exception:  # noqa
-        pass
-    secret_key = os.getenv("SECRET_KEY")
-    algorithm = os.getenv("JWT_ALGORITHM", "HS256")
-    user_obj = _build_jwt_payload(user.id)
-    token = jwt.encode(payload=user_obj, key=secret_key, algorithm=algorithm)
-    return {
-        "result": True,
-        "data": {
-            "id": user.id,
-            "username": user.username,
-            "display_name": user.display_name,
-            "is_first_login": is_first_login,
-            "locale": user.locale,
-            "timezone": user.timezone,
-            "token": token,
-        },
-    }
-
-
-def set_opspilot_guest_group_default_rule(default_group, user):
-    default_rule = GroupDataRule.objects.get(name="OpsPilot内置规则", app="opspilot", group_id=default_group.id)
-    monitor_rule = GroupDataRule.objects.get(name="OpsPilotGuest数据权限", app="monitor", group_id=default_group.id)
-    cmdb_rule = GroupDataRule.objects.get(name="游客数据权限", app="cmdb", group_id=default_group.id)
-    log_rule = GroupDataRule.objects.get(name="log内置规则", app="log", group_id=default_group.id)
-    node_rule = GroupDataRule.objects.get(name="节点管理内置数据权限", app="node", group_id=default_group.id)
-    UserRule.objects.get_or_create(username=user.username, group_rule_id=cmdb_rule.id)
-    UserRule.objects.get_or_create(username=user.username, group_rule_id=default_rule.id)
-    UserRule.objects.get_or_create(username=user.username, group_rule_id=monitor_rule.id)
-    UserRule.objects.get_or_create(username=user.username, group_rule_id=log_rule.id)
-    UserRule.objects.get_or_create(username=user.username, group_rule_id=node_rule.id)
-
-
-@nats_client.register
-def get_wechat_settings():
-    login_module = LoginModule.objects.filter(source_type="wechat", enabled=True).first()
-    if not login_module:
-        return {"result": True, "data": {"enabled": False}}
-
-    return {
-        "result": True,
-        "data": {
-            "enabled": True,
-            "app_id": login_module.app_id,
-            # app_secret 不再返回给前端，OAuth 验证已移至后端
-            "redirect_uri": login_module.other_config.get("redirect_uri", ""),
-            "callback_url": login_module.other_config.get("callback_url", ""),
-        },
-    }
-
-
-# generate_qr_code(username) 已于 #3439 废弃并移除：
-# 该 NATS handler 接受外部传入的 username，无任何调用方身份校验，
-# 任意内网服务均可通过单条消息覆盖任意用户的 OTP secret（认证凭据劫持）。
-# 安全替代接口请使用 generate_qr_code_by_user_id，
-# Web 层入口见 core/views/index_view.py::generate_qr_code。
-
-
-@nats_client.register
-def generate_qr_code_by_user_id(user_id):
-    """
-    Generate OTP QR code for a user by user_id.
-
-    This is the secure version that requires authentication.
-    The user_id comes from the authenticated session.
-    """
-    user = User.objects.filter(id=user_id).first()
-    if not user:
-        return {"result": False, "message": "User not found"}
-
-    user.otp_secret = pyotp.random_base32()
-    user.save()
-    totp = pyotp.TOTP(user.otp_secret)
-    # 创建用于Authenticator应用的配置URL
-    provisioning_uri = totp.provisioning_uri(name=user.username, issuer_name="WeopsX")
-
-    qr = qrcode.QRCode(
-        version=1,
-        error_correction=qrcode.constants.ERROR_CORRECT_L,
-        box_size=10,
-        border=4,
-    )
-    qr.add_data(provisioning_uri)
-    qr.make(fit=True)
-
-    img = qr.make_image(fill_color="black", back_color="white")
-    buffer = io.BytesIO()
-    img.save(buffer, format="PNG")
-    qr_code_base64 = base64.b64encode(buffer.getvalue()).decode()
-
-    return {"result": True, "data": {"qr_code": qr_code_base64}}
-
-
-# 验证OTP代码
-@nats_client.register
-def verify_otp_code(username, otp_code, client_ip=""):
-    """
-    Verify OTP code for a user by username.
-
-    Requires client_ip for rate limiting. Falls back to empty string when not
-    provided (legacy callers), but rate limiting is still applied per IP+username.
-    """
-    # Check rate limit before any DB lookup
-    is_limited, remaining = check_rate_limit(client_ip, username)
-    if is_limited:
-        return {"result": False, "message": "Too many failed attempts. Please try again later."}
-
-    user = User.objects.filter(username=username).first()
-    if not user:
-        return {"result": False, "message": "User not found"}
-
-    if not user.otp_secret:
-        return {"result": False, "message": "OTP not configured for this user"}
-
-    totp = pyotp.TOTP(user.otp_secret)
-    if totp.verify(otp_code):
-        reset_rate_limit(client_ip, username)
-        return {"result": True, "message": "Verification successful"}
-
-    record_failed_attempt(client_ip, username)
-    return {"result": False, "message": "Invalid OTP code"}
-
-
-@nats_client.register
-def verify_otp_code_by_user_id(user_id, otp_code):
-    """
-    Verify OTP code for a user by user_id.
-
-    This is the secure version that requires authentication.
-    The user_id comes from the authenticated session.
-    """
-    user = User.objects.filter(id=user_id).first()
-    if not user:
-        return {"result": False, "message": "User not found"}
-
-    if not user.otp_secret:
-        return {"result": False, "message": "OTP not configured for this user"}
-
-    totp = pyotp.TOTP(user.otp_secret)
-    if totp.verify(otp_code):
-        return {"result": True, "message": "Verification successful"}
-    return {"result": False, "message": "Invalid OTP code"}
-
-
-@nats_client.register
-def verify_otp_login(challenge_id, otp_code, client_ip=""):
-    """
-    Verify OTP code with challenge_id and issue JWT token.
-
-    This is the second phase of two-factor authentication.
-    The challenge_id was obtained from the first phase (password verification).
-
-    Args:
-        challenge_id: The challenge ID from password verification
-        otp_code: The OTP code from user's authenticator app
-        client_ip: Client IP for rate limiting
-
-    Returns:
-        Dict with login result and JWT token if successful
-    """
-    # Verify challenge exists and is valid
-    challenge_data = verify_challenge(challenge_id)
-    if not challenge_data:
-        return {"result": False, "message": "Invalid or expired challenge. Please login again."}
-
-    username = challenge_data.get("username")
-    user_id = challenge_data.get("user_id")
-
-    # Check rate limit
-    is_limited, remaining = check_rate_limit(client_ip, username)
-    if is_limited:
-        return {"result": False, "message": "Too many failed attempts. Please try again later."}
-
-    # Get user
-    user = User.objects.filter(id=user_id).first()
-    if not user:
-        return {"result": False, "message": "User not found"}
-
-    if not user.otp_secret:
-        return {"result": False, "message": "OTP not configured for this user"}
-
-    # Verify OTP code
-    totp = pyotp.TOTP(user.otp_secret)
-    if not totp.verify(otp_code):
-        record_failed_attempt(client_ip, username)
-        return {"result": False, "message": "Invalid OTP code"}
-
-    # OTP verified successfully
-    # Invalidate challenge (one-time use)
-    invalidate_challenge(challenge_id)
-
-    # Reset rate limit
-    reset_rate_limit(client_ip, username)
-
-    # Issue JWT token
-    secret_key = os.getenv("SECRET_KEY")
-    algorithm = os.getenv("JWT_ALGORITHM", "HS256")
-    user_obj = _build_jwt_payload(user.id)
-    token = jwt.encode(payload=user_obj, key=secret_key, algorithm=algorithm)
-
-    user.last_login = timezone.now()
-    user.save()
-
-    # Check password expiry reminder (expired users already blocked in login phase)
-    password_expiry_reminder = ""
-    if user.password_last_modified:
-        # Initialize language loader for user's locale
-        loader = LanguageLoader(app="system_mgmt", default_lang=user.locale or "en")
-
-        # 批量读取密码策略（单次 DB 查询 + 缓存，复用与 login() 相同的缓存键）
-        pwd_policy = _get_pwd_policy_settings()
-        validity_period_days = pwd_policy["pwd_set_validity_period"]
-
-        # validity_period_days <= 0 表示永不过期，跳过过期检查
-        if validity_period_days > 0:
-            reminder_days = pwd_policy["pwd_set_expiry_reminder_days"]
-
-            now = timezone.now()
-            password_expire_date = user.password_last_modified + timedelta(days=validity_period_days)
-            days_until_expire = (password_expire_date - now).days
-
-            if 0 < days_until_expire <= reminder_days:
-                password_expiry_reminder = loader.get(
-                    "login.password_expiring_soon",
-                    "Your password will expire in {days} day(s). Please change it soon.",
-                ).format(days=days_until_expire)
-
-    return {
-        "result": True,
-        "data": {
-            "token": token,
-            "username": username,
-            "display_name": user.display_name,
-            "id": user.id,
-            "domain": user.domain,
-            "locale": user.locale,
-            "timezone": user.timezone,
-            "temporary_pwd": user.temporary_pwd,
-            "password_expiry_reminder": password_expiry_reminder,
-        },
-    }
-
-
-@nats_client.register
-def get_namespace_by_domain(domain):
-    login_module = LoginModule.objects.filter(source_type="bk_lite", other_config__contains={"domain": domain}).first()
-    if not login_module:
-        return {"result": False, "message": "Login module not found"}
-    namespace = login_module.other_config.get("namespace", "")
-    return {"result": True, "data": namespace}
-
-
-@nats_client.register
-def bk_lite_user_login(username, domain):
-    user = User.objects.filter(username=username, domain=domain).first()
-    if not user:
-        return {"result": False, "message": "Username or password is incorrect"}
-    return get_user_login_token(user, username)
-
-
-def get_user_login_token(user, username, skip_token_for_otp=False):
-    """
-    Get login token for user.
-
-    Args:
-        user: User object
-        username: Username string
-        skip_token_for_otp: If True and OTP is enabled, return challenge_id instead of token
-
-    Returns:
-        Dict with login result and data
-    """
-    if user.disabled:
-        return {"result": False, "message": "User is disabled"}
-
-    # Check if OTP is enabled globally
-    enable_otp_setting = SystemSettings.objects.filter(key="enable_otp").first()
-    enable_otp = enable_otp_setting and enable_otp_setting.value == "1"
-
-    # Check if user has OTP configured (has otp_secret)
-    user_has_otp = user.otp_secret is not None and user.otp_secret != ""
-
-    # If OTP is enabled and we should use two-phase authentication
-    if skip_token_for_otp and enable_otp:
-        # Generate QR code for first-time OTP binding if user hasn't configured OTP yet
-        qr_code_base64 = None
-        if not user_has_otp:
-            # Generate new OTP secret and QR code
-            user.otp_secret = pyotp.random_base32()
-            user.save()
-            totp = pyotp.TOTP(user.otp_secret)
-            provisioning_uri = totp.provisioning_uri(name=username, issuer_name="WeopsX")
-
-            qr = qrcode.QRCode(
-                version=1,
-                error_correction=qrcode.constants.ERROR_CORRECT_L,
-                box_size=10,
-                border=4,
-            )
-            qr.add_data(provisioning_uri)
-            qr.make(fit=True)
-
-            img = qr.make_image(fill_color="black", back_color="white")
-            buffer = io.BytesIO()
-            img.save(buffer, format="PNG")
-            qr_code_base64 = base64.b64encode(buffer.getvalue()).decode()
-
-        challenge_id = create_challenge(user.id, username)
-        response_data = {
-            "require_otp": True,
-            "challenge_id": challenge_id,
-            "username": username,
-            "display_name": user.display_name,
-            "id": user.id,
-            "domain": user.domain,
-            "locale": user.locale,
-            "timezone": user.timezone,
-            "temporary_pwd": user.temporary_pwd,
-        }
-
-        # Include QR code for first-time binding
-        if qr_code_base64:
-            response_data["qr_code"] = qr_code_base64
-            response_data["need_bindng"] = True  # Flag to indicate first-time binding
-
-        return {
-            "result": True,
-            "data": response_data,
-        }
-
-    # Normal flow (OTP disabled): issue JWT token
-    secret_key = os.getenv("SECRET_KEY")
-    algorithm = os.getenv("JWT_ALGORITHM", "HS256")
-    user_obj = _build_jwt_payload(user.id)
-    token = jwt.encode(payload=user_obj, key=secret_key, algorithm=algorithm)
-    user.last_login = timezone.now()
-    user.save()
-
-    return {
-        "result": True,
-        "data": {
-            "token": token,
-            "username": username,
-            "display_name": user.display_name,
-            "id": user.id,
-            "domain": user.domain,
-            "locale": user.locale,
-            "timezone": user.timezone,
-            "temporary_pwd": user.temporary_pwd,
-            "enable_otp": enable_otp,
-            "qrcode": user.otp_secret is None or user.otp_secret == "",
-        },
-    }
-
-
-@nats_client.register
-def get_login_module_domain_list():
-    login_module_list = list(LoginModule.objects.filter(source_type="bk_lite").values_list("other_config__domain", flat=True))
-    login_module_list.insert(0, "domain.com")
-    return {"result": True, "data": login_module_list}
-
-
-@nats_client.register
-def delete_rules(group_ids, instance_id, app, module, child_module):
-    """
-    删除权限规则中指定实例的权限配置
-    """
-    try:
-        # 查询对应的 GroupDataRule
-        rules_queryset = GroupDataRule.objects.filter(group_id__in=group_ids, app=app)
-
-        updated_count = 0
-        affected_rule_ids = []
-        for rule_obj in rules_queryset:
-            rules_data = rule_obj.rules
-
-            # 如果没有对应的模块，跳过
-            if module not in rules_data:
-                continue
-
-            # 获取目标数据结构
-            if child_module:
-                # 二级模块，如 provider.llm_model
-                if child_module not in rules_data[module]:
-                    continue
-                target_list = rules_data[module][child_module]
-            else:
-                # 一级模块，如 skill、bot
-                target_list = rules_data[module]
-
-            # 删除指定 ID 的权限项
-            original_length = len(target_list)
-            if child_module:
-                rules_data[module][child_module] = [item for item in target_list if str(item.get("id")) != str(instance_id)]
-            else:
-                rules_data[module] = [item for item in target_list if str(item.get("id")) != str(instance_id)]
-
-            # 如果有删除操作，更新数据库
-            new_length = len(rules_data[module][child_module] if child_module else rules_data[module])
-            if new_length < original_length:
-                rule_obj.rules = rules_data
-                rule_obj.save()
-                updated_count += 1
-                affected_rule_ids.append(rule_obj.id)
-
-        # 清除受影响用户的权限缓存
-        if affected_rule_ids:
-            affected_users = list(UserRule.objects.filter(group_rule_id__in=affected_rule_ids).values("username", "domain"))
-            if affected_users:
-                clear_users_permission_cache(affected_users)
-
-        return {
-            "result": True,
-            "message": f"Successfully deleted rules from {updated_count} group data rules",
-        }
-
-    except Exception as e:
-        logger.exception(f"Error deleting rules: {e}")
-        return {"result": False, "message": str(e)}
-
-
-@nats_client.register
-def verify_bk_token(bk_token):
-    login_module = LoginModule.objects.filter(source_type="bk_login", enabled=True).first()
-    if not login_module:
-        return {"result": True, "data": {"bk_login_open": False}}
-    bk_config = login_module.other_config
-    if not bk_token:
-        return {
-            "result": True,
-            "data": {"bk_login_open": True, "user": {}, "url": bk_config.get("bk_url")},
-        }
-    res, bk_user = get_bk_user_info(
-        bk_token,
-        bk_config.get("app_id"),
-        bk_config.get("app_token"),
-        bk_config.get("bk_url"),
-    )
-    if not res:
-        return {
-            "result": True,
-            "data": {"bk_login_open": True, "user": {}, "url": bk_config.get("bk_url")},
-        }
-    group_obj = Group.objects.get(name=login_module.other_config.get("root_group", "蓝鲸"), parent_id=0)
-    user, _ = User.objects.get_or_create(
-        username=bk_user["username"],
-        domain=bk_user.get("domain"),
-        defaults={
-            "email": bk_user.get("email", ""),
-            "group_list": [group_obj.id],
-            "locale": bk_user.get("language", "zh-Hans"),
-            "timezone": bk_user.get("time_zone", "Asia/Shanghai"),
-            "role_list": login_module.other_config.get("default_roles", []),
-        },
-    )
-    user.email = bk_user.get("email", "")
-    user.locale = bk_user.get("language", user.locale)
-    user.timezone = bk_user.get("time_zone", user.timezone)
-    user.save()
-    user_obj = _build_jwt_payload(user.id)
-    secret_key = os.getenv("SECRET_KEY")
-    algorithm = os.getenv("JWT_ALGORITHM", "HS256")
-    token = jwt.encode(payload=user_obj, key=secret_key, algorithm=algorithm)
-    return {
-        "result": True,
-        "data": {
-            "bk_login_open": True,
-            "user": {
-                "token": token,
-                "username": user.username,
-                "display_name": user.display_name,
-                "id": user.id,
-                "domain": user.domain,
-                "locale": user.locale,
-                "timezone": user.timezone,
-                "qrcode": user.otp_secret is None or user.otp_secret == "",
-            },
-            "url": bk_config.get("bk_url"),
-        },
-    }
-
-
-@nats_client.register
-def save_error_log(username, app, module, error_message, domain="domain.com"):
-    """
-    保存错误日志
-    :param username: 用户名
-    :param app: 应用模块
-    :param module: 功能模块
-    :param error_message: 错误信息
-    :param domain: 域名
-    """
-    try:
-        ErrorLog.objects.create(
-            username=username,
-            app=app,
-            module=module,
-            error_message=error_message,
-            domain=domain,
-        )
-        return {"result": True, "message": "Error log saved successfully"}
-    except Exception as e:
-        logger.exception(f"Failed to save error log: {e}")
-        return {"result": False, "message": str(e)}
-
-
-@nats_client.register
-def save_operation_log(username, source_ip, app, action_type, summary="", domain="domain.com",
-                       target_type="", target_id="", detail=None):
-    """
-    保存操作日志
-    :param username: 用户名
-    :param source_ip: 源IP地址
-    :param app: 应用模块
-    :param action_type: 操作类型 (create/update/delete/execute)
-    :param summary: 操作概要
-    :param domain: 域名
-    :param target_type: 操作目标类型（可选）
-    :param target_id: 操作目标ID（可选）
-    :param detail: 操作详情 JSON（可选，默认空字典）
-    """
-    try:
-        # 验证 action_type 是否合法
-        valid_actions = [
-            OperationLog.ACTION_CREATE,
-            OperationLog.ACTION_UPDATE,
-            OperationLog.ACTION_DELETE,
-            OperationLog.ACTION_EXECUTE,
-        ]
-        if action_type not in valid_actions:
-            return {
-                "result": False,
-                "message": f"Invalid action_type. Must be one of: {', '.join(valid_actions)}",
-            }
-
-        OperationLog.objects.create(
-            username=username,
-            source_ip=source_ip,
-            app=app,
-            action_type=action_type,
-            summary=summary,
-            domain=domain,
-            target_type=target_type or "",
-            target_id=str(target_id or ""),
-            detail=detail or {},
-        )
-        return {"result": True, "message": "Operation log saved successfully"}
-    except Exception as e:
-        logger.exception(f"Failed to save operation log: {e}")
-        return {"result": False, "message": str(e)}
+def login(*args, **kwargs):
+    _sync_compat_globals()
+    return _login.login(*args, **kwargs)
